# Conflicts:
#	io.openems.edge.controller.symmetric.balancing/src/io/openems/edge/controller/symmetric/balancing/Balancing.java
#	io.openems.edge.ess.streetscooter/src/io/openems/edge/ess/streetscooter/AbstractEssStreetscooter.java

+++ io.openems.edge.ess.api/src/io/openems/edge/ess/dccharger/api/EssDcCharger.java
@@ -43,7 +43,7 @@ public interface EssDcCharger extends OpenemsComponent {
 				 * Fill Max Actual Power channel
 				 */
 				if (value.asOptional().isPresent()) {
-					int newValue = (int) value.get();
+					int newValue = (int) (Integer) value.get();
 					Channel<Integer> maxActualPowerChannel = channel.getComponent().channel(ChannelId.MAX_ACTUAL_POWER);
 					int maxActualPower = maxActualPowerChannel.value().orElse(0);
 					int maxNextActualPower = maxActualPowerChannel.getNextValue().orElse(0);
